Completed Version 1

+++ cv.c
@@ -12,42 +12,35 @@
 
 void cv_init(struct cv *cv){
 	cv->totalProcess=0;
-	char *cvaddr=mmap(NULL,(size_t)(sizeof(char)*64),PROT_READ|PROT_WRITE,MAP_SHARED|MAP_ANONYMOUS,-1,0);
-	cv->cvmutex=(struct spinlock *)cvaddr;
-	cv->cvmutex->primeLock=0;
+	int i=CV_MAXPROC;
+	for(i;i>=0;i--){
+		cv->pidArray[i]=0;
+	}
+	cv->cvmutex.primeLock=0;
 }
 
 void cv_wait(struct cv *cv, struct spinlock *mutex){
-	sigset_t sigmask;
-	if(sigfillset(&sigmask)<0){
-		fprintf(stderr, "Sig Set Error: %s\n", strerror(errno));
-		exit(EXIT_FAILURE);
-	}
-	if(sigdelset(&sigmask,10)<0){
-		fprintf(stderr, "Sig Delete Set Error: %s\n", strerror(errno));
-		exit(EXIT_FAILURE);
-	}
-	spin_lock(cv->cvmutex);
+	spin_lock(&cv->cvmutex);
 	cv->totalProcess++;
 	pid_t currentpid=getpid();
 	cv->pidArray[cv->totalProcess-1]=(int)currentpid;
-	printf("%d\n",currentpid);
-	spin_unlock(cv->cvmutex);
+	spin_unlock(&cv->cvmutex);
 	spin_unlock(mutex);
-	printf("$$$$\n");
-	sigsuspend(&sigmask);
+	sigsuspend(&cv->sigmask);
 	spin_lock(mutex);
 }
 int cv_broadcast(struct cv *cv){
-	spin_lock(cv->cvmutex);
+	spin_lock(&cv->cvmutex);
 	for(cv->totalProcess;cv->totalProcess>=0;cv->totalProcess--){
 		kill((pid_t)cv->pidArray[cv->totalProcess-1],SIGUSR1);
 	}
-	spin_unlock(cv->cvmutex);
+	spin_unlock(&cv->cvmutex);
 }
 int cv_signal(struct cv *cv){
-	spin_lock(cv->cvmutex);
-	cv->totalProcess--;
-	kill((pid_t)cv->pidArray[cv->totalProcess],SIGUSR1);
-	spin_unlock(cv->cvmutex);
+	spin_lock(&cv->cvmutex);
+	if(cv->totalProcess>0){
+		cv->totalProcess--;
+		kill((pid_t)cv->pidArray[cv->totalProcess],SIGUSR1);
+	}
+	spin_unlock(&cv->cvmutex);
 }
